+++ lib/pages/subscription/subscription_view_option.dart
@@ -1,0 +1,70 @@
+import 'package:flutter/material.dart';
+
+class SubscriptionOption extends StatelessWidget {
+  final bool isChecked;
+  final Function onchange;
+
+  const SubscriptionOption({
+    super.key,
+    required this.isChecked,
+    required this.onchange,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.all(20.0),
+      child: Container(
+        constraints: const BoxConstraints(maxWidth: 420),
+        padding: const EdgeInsets.only(
+          left: 16.0,
+          right: 16,
+          top: 8.0,
+          bottom: 8.0,
+        ),
+        decoration: BoxDecoration(
+          borderRadius: BorderRadius.circular(8.0),
+          border: Border.all(
+            color: Colors.grey,
+          ),
+          boxShadow: const [
+            BoxShadow(
+              color: Colors.black26,
+              offset: Offset(0, 2),
+            ),
+          ],
+        ),
+        child: Row(
+          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+          children: [
+            const Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                Text(
+                  'Yearly Subscription',
+                  style: TextStyle(
+                    fontSize: 18.0,
+                    fontWeight: FontWeight.bold,
+                  ),
+                ),
+                Text(
+                  'R 2,999',
+                  style: TextStyle(
+                    fontSize: 16.0,
+                    color: Colors.grey,
+                  ),
+                ),
+              ],
+            ),
+            Checkbox(
+              value: isChecked,
+              onChanged: (bool? value) {
+                onchange(value);
+              },
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
